+++ packages/docs/scripts/generators/component/Component.demo.vue.hbs
