+++ packages/docs/scripts/generators/component/Component.spec.ts.hbs
