+++ packages/docs/scripts/generators/component/Component.vue.hbs
