+++ packages/docs/scripts/generators/component/Example.vue.hbs
