+++ packages/docs/scripts/generators/component/api-options.ts.hbs
