+++ packages/docs/scripts/generators/component/page-config.ts.hbs
@@ -1,8 +1,12 @@
 import { ApiDocsBlock } from '../../../../types/configTypes'
+import Va{{properCase name}} from 'vuestic-ui/src/components/vuestic-components/va-{{kebabCase name}}/Va{{properCase name}}.vue'
+import apiOptions from './api-options'
 import { DocsHelper } from '../../../../helpers/DocsHelper'
 
 export default [
   DocsHelper.title('{{camelCase name}}.title'),
+
+  DocsHelper.subtitle('all.examples'),
 
   // examples
   // ...DocsHelper.exampleBlock(
@@ -10,4 +14,6 @@ export default [
   //   '{{camelCase name}}.examples.yourExample.text',
   //   'va-{{kebabCase name}}/Example',
   // ),
+
+  DocsHelper.api(Va{{properCase name}}, apiOptions),
 ] as ApiDocsBlock[]
