+++ packages/docs/scripts/generators/docspage/doc-page.vue.hbs
@@ -5,7 +5,7 @@
 <script lang="ts">
 import { Vue, Component } from 'vue-property-decorator'
 import DocsContent from '../../components/DocsContent.vue'
-import {{camelCase name}}Config from '../../components/page-configs/{{kebabCase category}}/va-{{kebabCase name}}/page-config'
+import {{camelCase name}}Config from '../../components/page-configs/{{kebabCase category}}/{{kebabCase name}}/page-config'
 
 @Component({
   components: {
