+++ packages/docs/scripts/generators/docspage/page-config.ts.hbs
@@ -1,12 +1,8 @@
 import { ApiDocsBlock } from '../../../../types/configTypes'
-import Va{{properCase name}} from 'vuestic-ui/src/components/vuestic-components/va-{{kebabCase name}}/Va{{properCase name}}.vue'
-import apiOptions from './api-options'
 import { DocsHelper } from '../../../../helpers/DocsHelper'
 
 export default [
   DocsHelper.title('{{camelCase name}}.title'),
-
-  DocsHelper.subtitle('all.examples'),
 
   // examples
   // ...DocsHelper.exampleBlock(
@@ -14,6 +10,4 @@ export default [
   //   '{{camelCase name}}.examples.yourExample.text',
   //   'va-{{kebabCase name}}/Example',
   // ),
-
-  DocsHelper.api(Va{{properCase name}}, apiOptions),
 ] as ApiDocsBlock[]
